codecoverage/bot: Make it possible to use HGMO class with a remote server

+++ src/codecoverage/bot/code_coverage_bot/hgmo.py
@@ -15,7 +15,13 @@ class HGMO(object):
     PID_FILE = 'hgmo.pid'
     SERVER_ADDRESS = 'http://localhost:8000'
 
-    def __init__(self, repo_dir):
+    def __init__(self, repo_dir=None, server_address=None):
+        assert (repo_dir is not None) ^ (server_address is not None)
+
+        if server_address is not None:
+            self.server_address = server_address
+        else:
+            self.server_address = HGMO.SERVER_ADDRESS
         self.repo_dir = repo_dir
         self.pid_file = os.path.join(os.getcwd(),
                                      HGMO.PID_FILE)
@@ -26,11 +32,15 @@ class HGMO(object):
             return int(pid)
 
     def __enter__(self):
+        if self.repo_dir is None:
+            return self
+
         proc = subprocess.Popen(['hg', 'serve',
                                  '--hgmo',
                                  '--daemon',
                                  '--pid-file', self.pid_file],
-                                cwd=self.repo_dir)
+                                cwd=self.repo_dir,
+                                stderr=subprocess.STDOUT)
         proc.wait()
 
         logger.info('hgmo is running', pid=self.__get_pid())
@@ -38,6 +48,9 @@ class HGMO(object):
         return self
 
     def __exit__(self, type, value, traceback):
+        if self.repo_dir is None:
+            return
+
         pid = self.__get_pid()
         os.killpg(os.getpgid(pid), signal.SIGTERM)
         os.remove(self.pid_file)
@@ -57,7 +70,7 @@ class HGMO(object):
         if changeset is not None:
             params['changeset'] = changeset
 
-        r = requests.get('{}/json-pushes'.format(HGMO.SERVER_ADDRESS), params=params)
+        r = requests.get('{}/json-pushes'.format(self.server_address), params=params)
 
         r.raise_for_status()
         return r.json()
@@ -69,7 +82,7 @@ class HGMO(object):
         return sum((data['changesets'] for data in push_data['pushes'].values()), [])
 
     def get_annotate(self, revision, path):
-        r = requests.get('{}/json-annotate/{}/{}'.format(HGMO.SERVER_ADDRESS, revision, path))
+        r = requests.get('{}/json-annotate/{}/{}'.format(self.server_address, revision, path))
 
         # 200 means success.
         # 404 means a file that doesn't exist (never existed or was removed).
